# BigInteger parsing optimizations