@danmosemsft 2021-02-05
> The use of ArrayPool<int> is much more common than ArrayPool<uint>

Grepped under libraries/*/src/**cs for interest and I see

object - 2
IntPtr, GCHandle - 4 each
char - 118
byte - 301

there are no uses I see for int or uint.