@tannergooding 2021-02-05
> Grepped under libraries/*/src/**cs for interest and I see

I wonder if there is something we can do here to promote better sharing. Particularly when using `Span` the underlying data type doesn't really matter too much (provided it has the correct alignment)