@jfd16 2021-02-05
> 
> 
> > The use of ArrayPool is much more common than ArrayPool
> 
> Grepped under libraries/*/src/**cs for interest and I see
> 
> object - 2
> IntPtr, GCHandle - 4 each
> char - 118
> byte - 301
> 
> there are no uses I see for int or uint.

`ArrayPool<int>` is being used indirectly in some places through `ValueListBuilder`: https://github.com/dotnet/runtime/search?q=ValueListBuilder+path%3Asrc