@danmoseley 2021-02-11
> I wonder if there is something we can do here to promote better sharing. Particularly when using Span the underlying data type doesn't really matter too much (provided it has the correct alignment

Worth an issue do you think @tannergooding